Move another function over to menu_driver.c
