+++ menu/menu_shader.h
@@ -182,6 +182,13 @@ void menu_driver_set_last_shader_path_int(
       char *shader_dir, size_t dir_len,
       char *shader_file, size_t file_len);
 
+bool dir_init_shader_internal(
+      bool shader_remember_last_dir,
+      struct rarch_dir_shader_list *dir_list,
+      const char *shader_dir,
+      const char *shader_file_name,
+      bool show_hidden_files);
+
 RETRO_END_DECLS
 
 #endif
